git-svn-id: https://gorgonlib.googlecode.com/svn/branches/2.x@464 a61f9e97-bd4c-0410-8c65-85b29eaa70ff

+++ Dorian/PlugIns/Gorgon.HID.RawInput/Devices/RawKeyboard.cs
@@ -184,7 +184,7 @@ namespace GorgonLibrary.HID.RawInput
 			System.Windows.Forms.Application.AddMessageFilter(_messageFilter);
 
 			_device.UsagePage = HIDUsagePage.Generic;
-			_device.Usage = HIDUsage.Keyboard;
+			_device.Usage = (ushort)HIDUsage.Keyboard;
 			_device.Flags = RawInputDeviceFlags.None;
 
 			// Enable background access.
@@ -214,7 +214,7 @@ namespace GorgonLibrary.HID.RawInput
 			}
 
 			_device.UsagePage = HIDUsagePage.Generic;
-			_device.Usage = HIDUsage.Keyboard;
+			_device.Usage = (ushort)HIDUsage.Keyboard;
 			_device.Flags = RawInputDeviceFlags.None;
 			_device.WindowHandle = IntPtr.Zero;
